# proviemplea/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProviEmplea",
        "description": "Documentación oficial de la API ProviEmplea.\n\n**Rate Limiting:** Máximo 60 solicitudes por minuto por IP.\n**Caché:** Los endpoints GET tienen caché de 5 minutos.\n**Tiempos de respuesta esperados:** GET ~100ms | POST ~200ms | PUT ~150ms | DELETE ~100ms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar personas activas",
                "description": "Retorna el listado de personas/talentos con estado activo.",
                "operationId": "19153d93fbbf375afea8d2a21f70831b",
                "parameters": [
                    {
                        "name": "modalidad",
                        "in": "query",
                        "description": "Filtrar por modalidad: Presencial, Remoto, Híbrido",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Remoto"
                        }
                    },
                    {
                        "name": "tipo_jornada",
                        "in": "query",
                        "description": "Filtrar por jornada: Completa, Part-time, Por proyecto",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Completa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Persona"
                                    }
                                },
                                "example": [
                                    {
                                        "id": "550e8400-e29b-41d4-a716-446655440000",
                                        "email": "talento01@email.com",
                                        "telefono": "987654321",
                                        "codigo_talento": "TAL-001",
                                        "nivel_educacional": "Universitario",
                                        "titulo_carrera": "Ingeniería Informática",
                                        "anios_experiencia": 3,
                                        "competencias": [
                                            "PHP",
                                            "Laravel",
                                            "MySQL"
                                        ],
                                        "tipo_jornada": "Completa",
                                        "modalidad": "Remoto",
                                        "validado": false,
                                        "activo": true
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Crear nueva persona/talento",
                "description": "Registra un nuevo talento. Solo se requieren email y teléfono.",
                "operationId": "b76ce2b01c2e821292fac2c4cddb74ee",
                "requestBody": {
                    "description": "Datos del talento a registrar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            },
                            "example": {
                                "email": "talento02@email.com",
                                "telefono": "912345678",
                                "codigo_talento": "TAL-002",
                                "nivel_educacional": "Técnico",
                                "titulo_carrera": "Técnico en Redes",
                                "anios_experiencia": 2,
                                "competencias": [
                                    "Redes",
                                    "Linux",
                                    "Soporte TI"
                                ],
                                "tipo_jornada": "Part-time",
                                "modalidad": "Híbrido"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener persona por ID",
                "description": "Retorna los datos de una persona usando su UUID.",
                "operationId": "9ae93a77b50e534e16e3a4aded57564d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la persona",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar persona completa (PUT)",
                "description": "Actualiza todos los campos de una persona existente.",
                "operationId": "a4370a9476fe98f54efa656fbddd395b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la persona",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            },
                            "example": {
                                "email": "actualizado@email.com",
                                "telefono": "911111111",
                                "nivel_educacional": "Postgrado",
                                "titulo_carrera": "Magíster en Ciencias de Datos",
                                "anios_experiencia": 6,
                                "competencias": [
                                    "Python",
                                    "Machine Learning",
                                    "SQL"
                                ],
                                "tipo_jornada": "Completa",
                                "modalidad": "Remoto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Eliminar persona",
                "description": "Borrado lógico: pone activo=false. La persona no aparece en listados pero sus datos se conservan.",
                "operationId": "65acb26d6a487dcc8cb74f199a297fe0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la persona",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Persona desactivada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/personas/{id}/toggle": {
            "patch": {
                "tags": [
                    "Personas"
                ],
                "summary": "Activar o desactivar perfil (PATCH)",
                "description": "Cambia el estado activo/inactivo de una persona. Operación de encender o apagar.",
                "operationId": "bc0a8dc17199be884e9d0b60e53e3c07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la persona",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado cambiado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estado actualizado correctamente"
                                        },
                                        "activo": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Persona": {
                "required": [
                    "email",
                    "telefono"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "talento01@email.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "987654321"
                    },
                    "codigo_talento": {
                        "type": "string",
                        "example": "TAL-001"
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "example": "Universitario"
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "example": "Ingeniería Informática"
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 3
                    },
                    "competencias": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "PHP"
                        }
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "example": "Completa"
                    },
                    "modalidad": {
                        "type": "string",
                        "example": "Remoto"
                    },
                    "validado": {
                        "type": "boolean",
                        "example": false
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}